requirements.txt
fastapi
uvicorn
scipy
numpy
pyyaml
pydantic